isolate get_log_file_list()
